Add Domains components
Add registration columns to application model.
Added check for default role for the application that can login.This way if the application is new or the user is new, we dont need to go and enable application login for that user. If the role, the user is in is allowed to login to that application, then we add permission during login attempt.