Added app_type DRYing code. Ecom->admin and Ecom->dashboard were both same application type (Dashboard) which means if we had more dashboards, we will have to duplicate code in each application. Also, because some packages were being shared across applications, we had to create a common package application which was just messy. Now, we can run multiple apps using the same app type. with different components.
Added exclusive_to_default_application setting in domain. This way, we can now tie a particular app to that domain. Which basically means that we dont need to add the route. Example if there is a domain serving eshop, we dont need to add /route/ to each link. $this->link will generate links as per the domain settings.